Feature/ui modify (#11)
* 添加修改密码的前端内容以及不完全的后台逻辑

* 修改密码功能前端部分（未完成）

* 前端修改密码功能（未完成）

* 完善了修改密码的后台逻辑

* 修改一个密码修改的逻辑错误

* 增加权限控制，非机构管理员无法创建项目或者审批

* 修改密码功能前端部分（未完成）

* 前端修改密码功能（未完成）

* 修改密码功能前端部分（未完成）

* 前端修改密码功能（未完成）

* 修改密码功能前端部分（未完成）

* 前端修改密码功能（未完成）

* 修改密码功能前端部分（未完成）

* 修改密码功能前端部分（未完成）

* 修改密码功能前端部分（未完成）

* 前端修改密码功能（未完成）

* 前端修改密码功能（未完成）

* 前端修改密码功能（未完成）

* init

* Add Account/Login

* 添加保存登录态功能。

* 修改上次提交。

* 准备解决csv逻辑问题，删除导入时的超过上限提醒

* 非管理员账户不显示审批列表

* 更改了时间显示格式

* 加入初始化时ping68进行网络检测的功能
csv加入try catch解决在分割字符串时出现的bug
修复了评分按钮和黑名单按钮不显示的bug

* 提交修改密码页面的更改

* 补全函数

* 添加了欢迎页。

* 修复批量添加中超过人数上限时弹窗过多问题；修复全选/反选按钮失效问题；修复处理冲突时由于点击过快可能出现的对某个Viemodel的nullexception；打开文件失败时不会再继续弹出批量处理的窗口

* UI美化：修改中文的登陆的label；新增按钮图标；注册页面布局调整、添加图标；ProjectDetail布局调整，引入StackPanel和Canvas的组合，按钮增加图标

* 增加图标

* 更改了按学号查找失败的错误

* 1.4.0.0 (#7)

* 添加修改密码的前端内容以及不完全的后台逻辑

* 修改密码功能前端部分（未完成）

* 前端修改密码功能（未完成）

* 完善了修改密码的后台逻辑

* 修改一个密码修改的逻辑错误

* 增加权限控制，非机构管理员无法创建项目或者审批

* 修改密码功能前端部分（未完成）

* 前端修改密码功能（未完成）

* 修改密码功能前端部分（未完成）

* 前端修改密码功能（未完成）

* 修改密码功能前端部分（未完成）

* 前端修改密码功能（未完成）

* 修改密码功能前端部分（未完成）

* 修改密码功能前端部分（未完成）

* 修改密码功能前端部分（未完成）

* 前端修改密码功能（未完成）

* 前端修改密码功能（未完成）

* 前端修改密码功能（未完成）

* Feature/ui modify (#6)

* init

* Add Account/Login

* 添加保存登录态功能。

* 修改上次提交。

* 准备解决csv逻辑问题，删除导入时的超过上限提醒

* 非管理员账户不显示审批列表

* 更改了时间显示格式

* 加入初始化时ping68进行网络检测的功能
csv加入try catch解决在分割字符串时出现的bug
修复了评分按钮和黑名单按钮不显示的bug

* 提交修改密码页面的更改

* 补全函数

* 添加了欢迎页。

* 修复批量添加中超过人数上限时弹窗过多问题；修复全选/反选按钮失效问题；修复处理冲突时由于点击过快可能出现的对某个Viemodel的nullexception；打开文件失败时不会再继续弹出批量处理的窗口

* UI美化：修改中文的登陆的label；新增按钮图标；注册页面布局调整、添加图标；ProjectDetail布局调整，引入StackPanel和Canvas的组合，按钮增加图标

* 增加图标

* 更改了按学号查找失败的错误

* [Fix]修复注册页关闭程序不关闭的bug。

* Develop (#9)

[Release] 1.4.0.0

* 添加修改密码的前端内容以及不完全的后台逻辑

* 修改密码功能前端部分（未完成）

* 前端修改密码功能（未完成）

* 完善了修改密码的后台逻辑

* 修改一个密码修改的逻辑错误

* 增加权限控制，非机构管理员无法创建项目或者审批

* Feature/ui modify (#6)

* init

* Add Account/Login

* 添加保存登录态功能。

* 修改上次提交。

* 准备解决csv逻辑问题，删除导入时的超过上限提醒

* 非管理员账户不显示审批列表

* 更改了时间显示格式

* 加入初始化时ping68进行网络检测的功能
csv加入try catch解决在分割字符串时出现的bug
修复了评分按钮和黑名单按钮不显示的bug

* 提交修改密码页面的更改

* 补全函数

* 添加了欢迎页。

* 修复批量添加中超过人数上限时弹窗过多问题；修复全选/反选按钮失效问题；修复处理冲突时由于点击过快可能出现的对某个Viemodel的nullexception；打开文件失败时不会再继续弹出批量处理的窗口

* UI美化：修改中文的登陆的label；新增按钮图标；注册页面布局调整、添加图标；ProjectDetail布局调整，引入StackPanel和Canvas的组合，按钮增加图标

* 增加图标

* 更改了按学号查找失败的错误

* [Fix]修复注册页关闭程序不关闭的bug。 (#8)

* Feature/ui modify (#6)

* init

* Add Account/Login

* 添加保存登录态功能。

* 修改上次提交。

* 准备解决csv逻辑问题，删除导入时的超过上限提醒

* 非管理员账户不显示审批列表

* 更改了时间显示格式

* 加入初始化时ping68进行网络检测的功能
csv加入try catch解决在分割字符串时出现的bug
修复了评分按钮和黑名单按钮不显示的bug

* 提交修改密码页面的更改

* 补全函数

* 添加了欢迎页。

* 修复批量添加中超过人数上限时弹窗过多问题；修复全选/反选按钮失效问题；修复处理冲突时由于点击过快可能出现的对某个Viemodel的nullexception；打开文件失败时不会再继续弹出批量处理的窗口

* UI美化：修改中文的登陆的label；新增按钮图标；注册页面布局调整、添加图标；ProjectDetail布局调整，引入StackPanel和Canvas的组合，按钮增加图标

* 增加图标

* 更改了按学号查找失败的错误

* [Fix]修复注册页关闭程序不关闭的bug。 (#8)

* 修正了前端的一些小错误。

* 调整了部分字体，改动了标签布局。

* 改进了显示缩放效果，继续调整界面布局。

* 补漏。

* 向处理冲突页面、确认添加页面、项目详情页面中添加了分页控件；修改了CsvHelper中异常处理问题

* 修复了主界面项目管理板块上，某些字段过长导致后方条目无法显示的问题。准备在项目内无志愿者时加入提示。

* 修复了评分不能使用的bug；修改了评分逻辑和UI；修改了增加评分时报错问题；增加了当项目管理者集合为空时的视觉提示；修复了各种DataGrid内容为空集时有一行空白的问题；更改了界面文字（英文->中文）；

* 修改了ProjectProgressHelper；修改了批量添加过程中综合评分显示为0的bug；完善了评分机制；对应地迁移了数据库

+++ UI of UIBE Volunteer Database/Pages/ProjectDetails.xaml.cs
@@ -21,6 +21,20 @@ namespace VolunteerDatabase.Desktop.Pages
     {
         private Project Pro;
         private AppUserIdentityClaims Claims;
+
+        private int VltCount { get { return Pro.Volunteers.Count; } }
+        private int VltPgeIndex { get; set; }
+        private int MaxVltItems = 5;
+
+        private int MngCount { get { return Pro.Managers.Count; } }
+        private int MngPgeIndex { get; set; }
+        private int MaxMngItems = 3;
+
+        public void finishScoringEventHandler()
+        {
+            ShowVltGrid();
+        }
+
         public void sendClaimsEventHandler(AppUserIdentityClaims claim)
         {
             IsEnabled = true;
@@ -46,21 +60,49 @@ namespace VolunteerDatabase.Desktop.Pages
                 this.Claims = Claim;
                 Pro = pro;
             }
+            //VltConut = Pro.Volunteers.Count;
             InitializeComponent();
-            Auth();        
+            Auth();
             ProInfoShow();
-            List<AppUser> users=Pro.Managers.ToList();
-            List<Volunteer> vols=Pro.Volunteers.ToList();
-            project_manager_list.ItemsSource = users;
-            volunteer_list.ItemsSource = vols;
+            ShowVltGrid();
+            ShowMngGrid();
+
         }
+
+        private void ShowVltGrid()
+        {
+            List<Volunteer> vols = new List<Volunteer>();
+            for (int i = VltPgeIndex; i <= VltPgeIndex + MaxVltItems; i++)
+            {
+                if (i > VltCount - 1) break;
+                vols.Add(Pro.Volunteers[i]);
+            }
+            VltPge.Content = string.Format("{0}/{1}",VltPgeIndex/MaxVltItems+1,VltCount/MaxVltItems+1);
+            LblVltListEpt.Visibility = Pro.Volunteers.Count == 0 ? Visibility.Visible : Visibility.Hidden;
+            volunteer_list.ItemsSource = vols;
+            //this.
+        }
+
+        private void ShowMngGrid()
+        {
+            List<AppUser> mngs = new List<AppUser>();
+            for (int i = MngPgeIndex; i <= MngPgeIndex + MaxMngItems; i++)
+            {
+                if (i > MngCount - 1) break;
+                mngs.Add(Pro.Managers[i]);
+            }
+            MngPge.Content = string.Format("{0}/{1}", MngPgeIndex / MaxMngItems + 1, MngCount / MaxMngItems + 1);
+            LblMngListEpt.Visibility =Pro.Managers.Count == 0?Visibility.Visible:Visibility.Hidden;
+            project_manager_list.ItemsSource = mngs;
+        }
+
         private void Auth()
         {
             if (Pro.ScoreCondition != ProjectScoreCondition.Scored)
             {
                 endproject.IsEnabled = false;
             }
-            if (Claims.Roles.Count() == 0||Pro.Condition==ProjectCondition.Finished)
+            if (Claims.Roles.Count() == 0 || Pro.Condition == ProjectCondition.Finished)
             {
                 AddManager_btn.IsEnabled = false;
                 deleteproject_btn.IsEnabled = false;
@@ -83,6 +125,7 @@ namespace VolunteerDatabase.Desktop.Pages
                 AddVolunteer_btn.IsEnabled = false;
             }
         }
+
         private void ProInfoShow()
         {
             if (Pro != null)
@@ -92,9 +135,9 @@ namespace VolunteerDatabase.Desktop.Pages
                 project_id.Text = Pro.Id.ToString();
                 project_place.Text = Pro.Place;
                 project_status.Text = Pro.Condition.ToString();
-                project_time.Text = Pro.Time.ToString();
+                project_time.Text = Pro.Time.ToString("yyyy-M-d");
                 project_accomodation.Text = Pro.Volunteers.Count() + "/" + Pro.Maximum.ToString();
-            }         
+            }
         }
 
         private void piliang_Click(object sender, RoutedEventArgs e)
@@ -110,102 +153,118 @@ namespace VolunteerDatabase.Desktop.Pages
                 {
                     foreach (string item in ch.informingMessage)
                     {
-                        //if(item)
                         MessageBox.Show(item);
-                    }//此处应建立窗口提示informingMessage,即有改动的信息，然后传多个学号，再调用ch中方法确定新信息
+                    }//显示非致命提示信息 此处应建立窗口提示informingMessage,即有改动的信息，然后传多个学号，再调用ch中方法确定新信息
                 }
-                if(ch.errorList.Count()!=0)
+                if (ch.errorList.Count() != 0)
                 {
                     foreach (string item in ch.errorList)
                     {
                         MessageBox.Show(item);
                     }
+                }//显示致命错误信息
+                else
+                {
+                    InputWindow window = new InputWindow();
+                    window.Height = 650;
+                    window.Width = 470;
+                    DealWithConflict dealer = new DealWithConflict(Pro, list, window);
+                    window.FinishInPutEvent += FinishInPutEventHandler;
+                    window.Content = dealer;
+                    window.Owner = this;
+                    window.Show();
                 }
-                InputWindow window = new InputWindow();
-
-                window.Height = 650;
-                window.Width = 470;
-                DealWithConflict dealer = new DealWithConflict(Pro, list, window);
-                window.FinishInPutEvent += FinishInPutEventHandler;
-                window.Content = dealer;
-                window.Owner = this;
-                window.Show();
                 //MessageBox.Show("导入的信息与志愿者库中不一致的条目已被红色高亮标记,请确认保留项目.");
             }
-
-          
-           
         }
 
         private void FinishInPutEventHandler()
         {
-            volunteer_list.ItemsSource = null;
-            volunteer_list.ItemsSource = Pro.Volunteers.ToList();
+            ShowVltGrid();
         }
 
         private void endproject_Click(object sender, RoutedEventArgs e)
         {
+            MessageBoxResult mr = MessageBox.Show("确认要结项吗？\n没有被评分的志愿者将被自动评分为4.","确认结项",MessageBoxButton.YesNo,MessageBoxImage.Question);
+            if(mr==MessageBoxResult.Yes)
             {
                 var pph = ProjectProgressHelper.GetInstance();
-                MessageBox.Show("未单独评分的志愿者将默认全部评分为：4");
-                if(Pro!=null&&Pro.Condition==ProjectCondition.Ongoing)
+                //MessageBox.Show("未单独评分的志愿者所有项全部评分为：4");
+                if (Pro != null && Pro.Condition == ProjectCondition.Ongoing)
                 {
-                    var result=pph.ScoringDefaultForVolunteers(Pro, 4);
-                    if(!result.Succeeded)
+                    try
                     {
-                        MessageBox.Show("评分失败");
+                        var result1 = pph.ScoringDefaultForVolunteers(Pro, new CreditRecord.CreditScore
+                        {
+                            CmmScore = 4,
+                            PncScore = 4,
+                            SrvScore = 4
+                        }
+                        );
+                        if (!result1.Succeeded)
+                        {
+                            MessageBox.Show("评分失败,未能结项\n"+string.Join(",",result1.Errors));
+                            return;
+                        }
+                        if (Pro.ScoreCondition == ProjectScoreCondition.Scored)
+                        {
+                            var result2 = pph.FinishProject(Pro);
+                            if (!result2.Succeeded)
+                            {
+                                MessageBox.Show("结项失败");
+                            }
+                        }
+                        else
+                        {
+                            MessageBox.Show("没有评分,未能结项");
+                            return;
+                        }
                     }
-                    
-                }          
-                if (Pro != null)
-                {
-                    var result = pph.FinishProject(Pro);
-                    if (!result.Succeeded)
+                    catch (Exception)
                     {
-                        MessageBox.Show("结项失败");
+                        throw;
                     }
-                }                        
+                }
             }
-
         }
+
         private void deleteproject_btn_Click(object sender, RoutedEventArgs e)
         {
             var pmh = ProjectManageHelper.GetInstance();
-            MessageBoxResult result=MessageBox.Show("确定要删除该项目?", "删除提醒", MessageBoxButton.YesNo, MessageBoxImage.Information);
-            switch(result)
+            MessageBoxResult result = MessageBox.Show("确定要删除该项目?", "删除提醒", MessageBoxButton.YesNo, MessageBoxImage.Information);
+            switch (result)
             {
                 case MessageBoxResult.Yes:
                     pmh.ProjectDelete(Pro);
-         
+
                     this.Close();
                     break;
+
                 case MessageBoxResult.No:
                     break;
             }
-
         }
 
         private void AddManager_btn_Click(object sender, RoutedEventArgs e)
         {
-            if(AddManager.Text=="")
+            if (AddManager.Text == "")
             {
                 MessageBox.Show("请输入管理者学号!");
             }
             else
-            { 
+            {
                 var pmh = ProjectManageHelper.GetInstance();
                 try
                 {
                     var result = pmh.AddManager(int.Parse(AddManager.Text), Pro);
-                    if(result.Succeeded)
+                    if (result.Succeeded)
                     {
-                        MessageBox.Show("学号为[" + AddManager.Text + "]的用户已经被添加为项目["+Pro.Name+"]的项目管理者.");
-                        project_manager_list.ItemsSource = null;
-                        project_manager_list.ItemsSource = Pro.Managers.ToList();
+                        MessageBox.Show("学号为[" + AddManager.Text + "]的用户已经被添加为项目[" + Pro.Name + "]的项目管理者.");
+                        ShowMngGrid();
                     }
                     if (!result.Succeeded)
                     {
-                        MessageBox.Show("导入失败,错误信息:"+string.Join(",",result.Errors));
+                        MessageBox.Show("导入失败,错误信息:" + string.Join(",", result.Errors));
                         AddManager.Clear();
                     }
                 }
@@ -218,9 +277,9 @@ namespace VolunteerDatabase.Desktop.Pages
 
         private void AddVolunteer_btn_Click(object sender, RoutedEventArgs e)
         {
-            if(AddVolunteer.Text=="")
+            if (AddVolunteer.Text == "")
             {
-                MessageBox.Show("请输入管理者学号!");
+                MessageBox.Show("请输入待添加的志愿者学号。");
             }
             else
             {
@@ -229,15 +288,14 @@ namespace VolunteerDatabase.Desktop.Pages
                 if (result.Succeeded)
                 {
                     MessageBox.Show("学号为[" + AddVolunteer.Text + "]的志愿者已经被添加入项目[" + Pro.Name + "]的志愿者列表.");
-                    volunteer_list.ItemsSource = null;
-                    volunteer_list.ItemsSource = Pro.Volunteers.ToList();
+                    ShowVltGrid();
                 }
                 if (!result.Succeeded)
                 {
-                    MessageBox.Show("导入失败");
+                    MessageBox.Show("导入失败，"+string.Join(",",result.Errors));
                     AddVolunteer.Clear();
                 }
-            }          
+            }
         }
 
         private void AddManager_KeyDown(object sender, KeyEventArgs e)
@@ -256,19 +314,9 @@ namespace VolunteerDatabase.Desktop.Pages
             }
         }
 
-        private void rate_btn_Click(object sender, RoutedEventArgs e)
+        /*private void rate_btn_Click(object sender, RoutedEventArgs e)
         {
-            Button senderButton = sender as Button;
-            if (senderButton.DataContext is Volunteer)
-            {
-                Volunteer Vol = (Volunteer)senderButton.DataContext;
-                if(Vol!=null&&Pro.ScoreCondition!=ProjectScoreCondition.UnScored)
-                {
-                    var rp = new Rating(Vol,Pro);
-                    rp.Owner = this;
-                }
-            }          
-        }
+        }*/
 
         private void deleteprojectmanager_btn_Click(object sender, RoutedEventArgs e)
         {
@@ -276,12 +324,11 @@ namespace VolunteerDatabase.Desktop.Pages
             if (senderButton.DataContext is AppUser)
             {
                 AppUser Man = (AppUser)senderButton.DataContext;
-                if(Man!=null)
+                if (Man != null)
                 {
                     var pmh = ProjectManageHelper.GetInstance();
                     pmh.DeletManager(Man.StudentNum, Pro);
-                    project_manager_list.ItemsSource = null;
-                    project_manager_list.ItemsSource = Pro.Managers.ToList();
+                    ShowMngGrid();
                 }
             }
         }
@@ -294,10 +341,13 @@ namespace VolunteerDatabase.Desktop.Pages
                 Volunteer Vol = (Volunteer)senderButton.DataContext;
                 if (Vol != null)
                 {
-                    var pph = ProjectProgressHelper.GetInstance();
-                    pph.DeleteVolunteerFromProject(Vol, Pro);
-                    volunteer_list.ItemsSource = null;
-                    volunteer_list.ItemsSource = Pro.Volunteers.ToList();
+                    MessageBoxResult mr = MessageBox.Show(string.Format("确认从项目中移除姓名为[{0}],学号为[{1}]的志愿者？", Vol.Name, Vol.StudentNum), "确认移除", MessageBoxButton.YesNo);
+                    if (mr == MessageBoxResult.Yes)
+                    {
+                        var pph = ProjectProgressHelper.GetInstance();
+                        pph.DeleteVolunteerFromProject(Vol, Pro);
+                        ShowVltGrid();
+                    }
                 }
             }
         }
@@ -332,5 +382,84 @@ namespace VolunteerDatabase.Desktop.Pages
                 MessageBox.Show("用户取消操作。");
             }
         }
+
+        private void BlacklistDetails_btn_Click(object sender, RoutedEventArgs e)
+        {
+            Button senderButton = sender as Button;
+            if (senderButton.DataContext is Volunteer)
+            {
+                Volunteer Vol = (Volunteer)senderButton.DataContext;
+                if (Vol != null)
+                {
+                    var bl = new BlackList(Vol);
+                    bl.Show();
+                    bl.Owner = this;
+                }
+            }
+        }
+
+        private void VltPgeNext_Click(object sender, RoutedEventArgs e)
+        {
+            VltPgeIndex = (VltPgeIndex + MaxVltItems > VltCount) ? VltPgeIndex : VltPgeIndex + MaxVltItems;
+            if (VltPgeIndex + MaxVltItems > VltCount) VltPgeNext.IsEnabled = false;
+            VltPgePrevious.IsEnabled = true;
+            ShowVltGrid();
+        }
+
+        private void VltPgePrevious_Click(object sender, RoutedEventArgs e)
+        {
+            VltPgeIndex = (VltPgeIndex - MaxVltItems < 0) ? VltPgeIndex : VltPgeIndex - MaxVltItems;
+            if (VltPgeIndex - MaxVltItems < 0) VltPgePrevious.IsEnabled = false;
+            VltPgeNext.IsEnabled = true;
+            ShowVltGrid();
+        }
+
+        private void MngPgeNext_Click(object sender, RoutedEventArgs e)
+        {
+            MngPgeIndex = (MngPgeIndex + MaxMngItems > MngCount) ? MngPgeIndex : MngPgeIndex + MaxMngItems;
+            if (MngPgeIndex + MaxMngItems > MngCount) MngPgeNext.IsEnabled = false;
+            MngPgePrevious.IsEnabled = true;
+            ShowMngGrid();
+        }
+
+        private void MngPgePrevious_Click(object sender, RoutedEventArgs e)
+        {
+            MngPgeIndex = (MngPgeIndex - MaxMngItems < 0) ? MngPgeIndex : MngPgeIndex - MaxMngItems;
+            if (MngPgeIndex - MaxMngItems < 0) MngPgePrevious.IsEnabled = false;
+            MngPgeNext.IsEnabled = true;
+            ShowMngGrid();
+        }
+
+
+
+        private void BtnRate_Click(object sender, RoutedEventArgs e)
+        {
+            Button senderButton = sender as Button;
+            if (senderButton.DataContext is Volunteer)
+            {
+                Volunteer Vol = (Volunteer)senderButton.DataContext;
+                if (Vol != null && Pro.ScoreCondition == ProjectScoreCondition.UnScored)
+                {
+                    var rp = new Rating(Vol, Pro,finishScoringEventHandler);
+                    rp.Owner = this;
+                    rp.Show();
+                }
+            }
+            else
+            {
+                MessageBox.Show("数据上下文非志愿者，请联系管理员。");
+            }
+        }
+        
+        //public void Query(int size, int pageIndex)
+        //{
+        //    Result.Total = Student.Students.Count;
+        //    Result.Students = Student.Students.Skip((pageIndex - 1) * size).Take(size).ToList();
+
+        //}
+        //private void dataPager_PageChanged(object sender, Resource.PageChangedEventArgs args)
+        //{
+        //    Query(args.PageSize, args.PageIndex);
+        //}
     }
 }
